fix: 🐛 keep pino out of the eve agent bundle
